// src/interfaces/filterEntities.interface.ts
export interface IAvailability {
  available: boolean;
  rented: boolean;
}

export interface ILanguages {
  english: boolean;
  spanish: boolean;
  french: boolean;
}

export interface IGenres {
  fiction: boolean;
  nonFiction: boolean;
  mystery: boolean;
  fantasy: boolean;
  romance: boolean;
}